Curious_PM/requirements.txt
streamlit==1.24.0
moviepy==1.0.3
google-cloud-speech==2.17.1
google-cloud-texttospeech==2.17.1
openai==0.27.8
pillow==9.5.0
